T-11741 facetoface: Add sessions reservations by managers
This is the face-to-face module enhancement which was ordered by xtractor and delivered by synergy learning. It adds
some pretty cool features including the ability for managers and admins to "reserve" spaces in sessions for users
without naming them specifically (which can be done later or not at all). These spaces can be automatically removed
before the session start time, etc.

    The main purpose(s) with the new extension is:
    A)  To support un-named reservations by a manager (or admin) for a specific F2F-session
    B)  To support the workflow where a staff manager signs up members of his team to a specific F2F-session

    As you know these are common features of other corporate LMS (Saba) and we think they would be welcomed by the
community.

Change-Id: I00e1f0683488388c819841aeda0a9f158a4d394c
Reviewed-on: https://review.totaralms.com/4601
Reviewed-by: Alastair Munro <alastair.munro@totaralms.com>
Reviewed-by: Maria Torres <maria.torres@totaralms.com>
Tested-by: Simon Coggins <simon.coggins@totaralms.com>