T-12892 facetoface: Add cut-off date / Minimum Capacity to sessions
This is a new setting to sessions that defines the minimum number of seats for a session. This will be a checkbox (default unchecked). If checked then the setting below becomes enabled.

A new cut off date selector setting will also be configured. The cut off date can't be on the same date as the planned session start date (if defined) and must be before it.

On the cut of date any user with course editing capability will receive a notification to indicate that the session is under capacity if it is so.

Thanks to Synergy Learning for the code.

Change-Id: I879196364ff9fab59c9ec6f3139d6a63a905b0b4
Reviewed-on: https://review.totaralms.com/5520
Tested-by: Jenkins Automation <jenkins@totaralms.com>
Reviewed-by: Simon Coggins <simon.coggins@totaralms.com>