T-12900 facetoface: Add self-approval feature
This is an optional extension to the face to face sign up form allowing the user to self certify they have sought approval for signing up to the session.

A new checkbox setting for the session setting to turn this feature on or off. This will only be enabled if approval is required for the activity itself.

When configured, on the sign up form a new checkbox will be shown that MUST be checked (unchecked initially) before submission is allowed. In this case the booking is approved straight away.

A link to the T&C’s for sign up is shown (popup window). The T&C's can be configured in the activity settings

Thanks to Synergy Learning for contributing this feature.

Change-Id: I22e8af89fb3ff60a5bc7956e97040079dd6005ce
Reviewed-on: https://review.totaralms.com/5521
Reviewed-by: Maria Torres <maria.torres@totaralms.com>
Reviewed-by: Simon Coggins <simon.coggins@totaralms.com>
Tested-by: Jenkins Automation <jenkins@totaralms.com>