T-12902 facetoface: Add declare interest feature
Similar to the concept of waiting list this allows users to be added to a list of users who be interested in the seminar but the session list hasn't be released yet or when there is no spaces available on any sessions but they would like to be considered for the future.

Therefore the user needs to be able to declare interest at the activity level and not the session level.

Thanks to Synergy Learning for developing this feature

Change-Id: If2af37d10f24d5f507d931d6260952c5ec0233f1
Reviewed-on: https://review.totaralms.com/5522
Tested-by: Jenkins Automation <jenkins@totaralms.com>
Reviewed-by: Maria Torres <maria.torres@totaralms.com>
Reviewed-by: Simon Coggins <simon.coggins@totaralms.com>