TL-8383 mod: do not recalculate completion immediately when unlocking and deleting mod completion
This is a change in behaviour, previously if you unlocked and deleted activity completion then activity completion would be
reaggregated for the activity for all users immediately.
This could lead to course + prog + certification reaggregation as well and take an incredibly long time.
As of this patch that reaggregation is done on cron by the \core\task\completion_regular_task cron task rather than immediately.
This is consistent with courses, and certifications.

Change-Id: I403fe4b8186b7d909c712715d1ab35fa1d2fe951
Reviewed-on: https://review.totaralms.com/10909
Reviewed-by: Nathan Lewis <nathan.lewis@totaralms.com>
Reviewed-by: Sam Hemelryk <sam.hemelryk@totaralms.com>
Tested-by: Jenkins Automation <jenkins@totaralms.com>