TL-15781 / MDL-49398 accesslib: Simplify role definition loading and cache
This change removes the complex and partial loading of role definition data for users.
It is replaced by keeping one system-wide definition for each role in MUC.

This is cheaper to regenerate as it avoids complex database queries (repeated LEFT JOIN to context).
Memory usage is kept low thanks to array CoW for each user access data.

Change-Id: I0257ef1467fd17be4ff6a242370b1380a02f52f4
Reviewed-on: https://review.totaralms.com/16479
Reviewed-by: Sam Hemelryk <sam.hemelryk@totaralearning.com>
Tested-by: Sam Hemelryk <sam.hemelryk@totaralearning.com>