TL-11448 core: fixed flex icon delivery through web services
A change in 3.1 meant the structure of the flex icons data via
webservices was no longer valid, nor could it be validated.
As a hack until we get time to improve the situation we use a
dedicated class for the flex icon return value that skips
the validation of the response.