TL-12399 core_block: Add option for borderless blocks
Make it possible for a block to be configured without any outer border.
This is useful for blocks which provide their own outer border to avoid
double borders. Each block type can specify if they should be borderless
by default, but this can be overridden on a per-instance basis in the
block configuration.

Change-Id: Iefcc3bbdc05484f4d50442fd10d9f8a0d8f03bc5
Reviewed-on: https://review.totaralms.com/14373
Reviewed-by: Sam Hemelryk <sam.hemelryk@totaralearning.com>
Tested-by: Sam Hemelryk <sam.hemelryk@totaralearning.com>