TL-15635 / MDL-51932 workshop: Improve UX for teachers on switching workshop phases
The icon itself has been found insufficient, teachers did not realize it
was used to switch the phase. Its semantics is even less intuitive in
Boost when we do not show it on course outline any more.

So this patch starts to display the information that has been present
for assistive technologies only. As a side effect, I spotted there was a
bug that caused the id attribute "mod_workshop-userplancurrenttasks"
was used for all titles (not only the current one) so it did not
really serve the purpose well.