TL-22677 / MDL-60209 mathjax: Do not load broken accessibility extensions
The Accessible.js config loads accessibility-menu extension which in
turn loads mathjax-sre.js library. There is a bug in this library
https://github.com/zorkow/speech-rule-engine/issues/182 that makes it
fail in RequireJS environments. This has been reported upstream as
https://github.com/mathjax/MathJax/issues/1854

As an immediate solution for now, this patch changes the default MathJax
configuration so that it does not load accessibility extensions at all.